# openapi.yaml
openapi: 3.0.3
info:
  title: E-commerce Graphic Designer API
  description: AI-powered image editing service for e-commerce product images
  version: 1.0.0
  contact:
    name: Development Team
    email: dev@example.com
servers:
  - url: http://localhost:3001/api
    description: Development server
  - url: https://api.example.com/api
    description: Production server
paths:
...
  /edit:
    post:
      tags: ['AI Processing']
      summary: Create image processing job
      description: Submit an image for AI processing (optimize, edit, or refine)
      parameters:
        - name: idempotency-key
          in: header
          required: false
          description: Idempotency key to prevent duplicate processing
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - session_id
                - image_id
                - type
              properties:
                session_id:
                  type: string
                  format: uuid
                image_id:
                  type: string
                  format: uuid
                type:
                  type: string
                  enum: [optimize, edit, refine]
                  description: Type of processing to perform
                prompt:
                  type: string
                  description: Processing instructions or description
                model:
                  $ref: '#/components/schemas/AIModelType'
                  description: Preferred AI model (optional, will use recommended if not specified)
      responses:
        '200':
          description: Job created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          job_id:
                            type: string
                            format: uuid
                          model:
                            $ref: '#/components/schemas/AIModelType'
                            description: The model that will be used for processing
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /edit/refine:
    post:
      tags: ['AI Processing']
      summary: Refine an existing variant
      description: Create a refinement job based on an existing variant
      parameters:
        - name: idempotency-key
          in: header
          required: false
          description: Idempotency key to prevent duplicate processing
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - session_id
                - variant_id
                - instructions
              properties:
                session_id:
                  type: string
                  format: uuid
                variant_id:
                  type: string
                  format: uuid
                instructions:
                  type: string
                  description: Refinement instructions
      responses:
        '200':
          description: Refinement job created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          job_id:
                            type: string
                            format: uuid
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /edit/models:
    get:
      tags: ['AI Processing']
      summary: Get available AI models
      description: Retrieve list of available AI models and their capabilities
      responses:
        '200':
          description: Available models information
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          models:
                            type: array
                            items:
                              $ref: '#/components/schemas/AIModelInfo'
                          default:
                            $ref: '#/components/schemas/AIModelType'
                          total:
                            type: integer
        '500':
          $ref: '#/components/responses/InternalError'
  /edit/models/{model}/test:
    post:
      tags: ['AI Processing']
      summary: Test AI model connection
      description: Test connectivity and availability of a specific AI model
      parameters:
        - name: model
          in: path
          required: true
          description: Model to test
          schema:
            $ref: '#/components/schemas/AIModelType'
      responses:
        '200':
          description: Connection test results
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          model:
                            $ref: '#/components/schemas/AIModelType'
                          connection:
                            type: object
                            properties:
                              success:
                                type: boolean
                              error:
                                type: string
                                nullable: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  schemas:
    ApiResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          description: Whether the request was successful
        data:
          type: object
          nullable: true
          description: Response data (present on success)
        error:
          type: string
          nullable: true
          description: Error message (present on failure)
    AIModelType:
      type: string
      enum: [gemini, chatgpt, sora]
      description: Available AI models
    AIModelInfo:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/AIModelType'
        name:
          type: string
          description: Human-readable model name
        description:
          type: string
          description: Model description
        capabilities:
          type: array
          items:
            type: string
            enum: [optimize, edit, refine, background_replace]
          description: Supported operations
        tier:
          type: string
          enum: [premium, creative, standard]
          description: Model quality tier
        available:
          type: boolean
          description: Whether the model is currently available
...
  responses:
    BadRequest:
      description: Bad request - invalid input parameters
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Invalid request parameters"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Resource not found"
    RateLimited:
      description: Too many requests - rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
          description: Number of seconds to wait before retrying
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Rate limit exceeded. Please retry after specified time."
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Internal server error"
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication (if implemented)
...